fix api mistake
