+++ backend/api/thermo.py
@@ -5,25 +5,25 @@ from fastapi import APIRouter, Depends, HTTPException
 
 from backend.api.db import get_db
 from backend.database import Database
-from backend.models.kinetics import Kinetics
+from backend.models.thermo import Thermo
 
 router = APIRouter()
 
 
-@router.post("/kinetics", response_model=Kinetics)
-def create_kinetics(kinetics: Kinetics, db: Database = Depends(get_db)) -> Kinetics:
-    db.import_kinetics(kinetics)
-    return kinetics
+@router.post("/thermo", response_model=Thermo)
+def create_thermo(thermo: Thermo, db: Database = Depends(get_db)) -> Thermo:
+    db.import_thermo(thermo)
+    return thermo
 
 
-@router.get("/kinetics/{kinetics_id}", response_model=Kinetics)
-def get_kinetics(kinetics_id: UUID, db: Database = Depends(get_db)) -> Kinetics:
+@router.get("/thermo/{thermo_id}", response_model=Thermo)
+def get_thermo(thermo_id: UUID, db: Database = Depends(get_db)) -> Thermo:
     try:
-        return db.get_kinetics(kinetics_id)
+        return db.get_thermo(thermo_id)
     except KeyError:
         raise HTTPException(status_code=404, detail="Model not found")
 
 
-@router.get("/kinetics", response_model=Iterable[Kinetics])
-def get_kineticss(db: Database = Depends(get_db)) -> Iterable[Kinetics]:
-    return db.get_all_kinetics()
+@router.get("/thermo", response_model=Iterable[Thermo])
+def get_thermos(db: Database = Depends(get_db)) -> Iterable[Thermo]:
+    return db.get_all_thermo()
